New module kas-core::event::cx::window
